backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-doc-analyzer"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.3.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.8.0",
    "openai>=1.30.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "spacy>=3.7.0",
    "chardet>=5.2.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "slowapi>=0.1.9",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    "pgvector>=0.3.0",
    "python-dateutil>=2.9.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
